% Generated by build script
\name{berk_global_temp}
\alias{berk_global_temp}
\title{Fetch Berkeley Earth global land+ocean temperature data}
\usage{
berk_global_temp()
}
\value{tibble: year (integer), month (integer), anomaly (numeric),}
\description{Fetch Berkeley Earth global land+ocean temperature data}
